Upgrade to new Sidekick Plugin Installer Version
