+++ analysis_options.yaml
@@ -3,3 +3,4 @@ include: package:lint/analysis_options.yaml
 linter:
   rules:
     avoid_print: false
+    sort_pub_dependencies: false
