+++ pubspec.yaml
@@ -12,5 +12,10 @@ dependencies:
   sidekick_core: '>=0.7.1 <1.0.0'
 
 dev_dependencies:
-  lint: ^1.5.0
-  sidekick_plugin_installer: ^0.1.3
+  lint: ^2.0.0
+  sidekick_plugin_installer: ^0.2.1
+
+  # Template Dependencies
+  shelf_static: ^1.1.0
+  shelf_router: ^1.1.3
+  shelf_helmet: ^1.1.0
