+++ template/pubspec.template.yaml
@@ -1,0 +1,15 @@
+name: dockerize_server
+description: A new Dockerized Flutter App
+
+
+publish_to: 'none'
+
+version: 1.0.0+1
+
+environment:
+  sdk: ">=2.15.0 <3.0.0"
+
+dependencies:
+  shelf_static: ^1.1.0
+  shelf_router: ^1.1.3
+  shelf_helmet: ^1.1.0
